refactor: introduced suspense/fallback mechanism. Fixed SSR errors for admin dashboard.
